Create Book Application Form(JDBC)
Design a Book application form using HTML/JSP. Have input fields for title, author, number of pages and price of the book. 
There are two buttons SAVE and FIND along with a CLEAR button. When SAVE button is clicked, the form details must be stored to the database (servlet-jdbc). When FIND button is clicked, with only the title of the book entered, retrieve the other details of the book (if available) from the database and display on the screen. Otherwise display an error message and provide a link to the main form.